file: add read()/write() operations
File must be opened in read/write mode. This is based on std::ios
standard mode flags.

This is based on the fstream implementation, which is buffered on some
systems.
flush() can be used to synchronize back to hardware.
http://en.cppreference.com/w/cpp/io/basic_ostream/flush

An alternative would have been boost::asio, which is probably better.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>