file: reworked to remove fstream member
When working on the FileSystem class, i noticed that fstreams are not
copyable. This is a problem since i need to be able to copy Files to put
them into my FileSystem collection.

This patch reworks the File class to no longer have a fstream member,
which makes the File class copyable.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>